requirements.txt
arabic-reshaper==2.1.3
asgiref==3.8.1
astroid==2.4.2
auditlog3==1.0.1
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2020.12.5
cffi==1.14.4
chainmap==1.0.3
chardet==4.0.0
click==8.1.2
combomethod==1.0.12
cryptography==3.3.1
csscompressor==0.9.5
defusedxml==0.6.0
device-detector==5.0.1
diff-match-patch==20200713
django==4.2.21
django-ajax-selects==2.2.0
django-allauth==0.47.0
django-blacklist==0.7.0
django-common-helpers==0.9.2
django-commonstuff==0.9
django-cors-headers==3.7.0
django-crispy-forms==1.11.0
django-cron==0.5.1
django-debug-toolbar==3.2.4
django-email-log==1.0.0
django-infopush==1.8
django-ipware==4.0.2
django-mailbox==4.8.2
django-user-agents==0.4.0
django-utils-six==2.0
django-xhtml2pdf==0.0.4
djangorestframework==3.15.2
drf-nested-routers==0.94.2
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
http-ece==1.1.0
idna==2.10
isort==5.7.0
jsmin==2.2.2
lazy-object-proxy==1.4.3
lru-cache==0.2.3
MarkupPy==1.14
mccabe==0.6.1
msal==1.22.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nulltype==2.3.1
numpy==1.24.2
oauthlib==3.1.0
odfpy==1.4.1
Office365-REST-Python-Client==2.4.0
openpyxl==3.0.10
options==1.4.10
packaging==23.1
pandas==2.0.0
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.5.1
plotly==5.14.1
ppjson==0.0.1
pprintjson==1.4.2
protobuf==3.19.3
py-vapid==1.7.1
pycparser==2.20
pycrypto==2.6.1
Pygments==2.11.2
PyJWT==2.0.1
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
PyMySQL==1.0.3
pyngrok==5.2.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.1
python3-openid==3.2.0
pytils==0.3
pytz==2021.1
pywebpush==1.11.0
PyYAML==5.4.1
regex==2022.1.18
reportlab==3.6.5
requests==2.25.1
requests-oauthlib==1.3.0
robot-detection==0.4
six==1.15.0
sly==0.4
sqlparse==0.4.1
tablib==3.3.0
tenacity==8.2.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
tzdata==2023.3
ua-parser==0.10.0
urllib3==1.24.2
user-agents==2.2.0
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
